requirements.txt
docutils
pyyaml
transitions
pytest
# testrail
